creating individual team member profiles
